+++ apps/@examples/embedded-contextual-education/src/pages/_app.tsx
@@ -32,7 +32,7 @@ export function App() {
     <DoptProvider
       apiKey={import.meta.env.VITE_DOPT_BLOCKS_API_KEY}
       userId={userId}
-      flowVersions={{ 'embedded-contextual-education2': 1 }}
+      flows={{ 'embedded-contextual-education2': 1 }}
     >
       <ChakraProvider>
         <Page />