+++ components/@dopt/react/card/sandbox/src/pages/_app.tsx
@@ -7,7 +7,7 @@ export function App() {
     <DoptProvider
       userId="joe_mckenney"
       apiKey="blocks-blocksKey_Mg=="
-      flowVersions={{
+      flows={{
         card: 0,
       }}
     >