+++ components/@dopt/react/help-hub/src/help-hub/launcher.tsx
@@ -24,25 +24,6 @@ export interface LauncherProps
   isOpen: boolean;
 }
 
-const HelpIcon = () => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="44"
-    height="44"
-    viewBox="0 0 24 24"
-    strokeWidth="1.5"
-    stroke="currentColor"
-    fill="none"
-    strokeLinecap="round"
-    strokeLinejoin="round"
-  >
-    <path stroke="none" d="M0 0h24v24H0z" fill="none" />
-    <path d="M3 12a9 9 0 1 0 18 0a9 9 0 0 0 -18 0" />
-    <path d="M12 16v.01" />
-    <path d="M12 13a2 2 0 0 0 .914 -3.782a1.98 1.98 0 0 0 -2.414 .483" />
-  </svg>
-);
-
 const footerClassName = `${classNameRoot}__launcher` as const;
 
 function HelpHubLauncher(
@@ -64,7 +45,7 @@ function HelpHubLauncher(
       className={clsx([
         themeClassName({
           theme,
-          className: classes.helpHubLauncher({ open: isOpen }),
+          className: classes.helpHubLauncher,
         }),
         footerClassName,
         className,
@@ -73,7 +54,25 @@ function HelpHubLauncher(
       {...restProps}
       ref={ref}
     >
-      <HelpIcon />
+      <svg
+        className={clsx([
+          themeClassName({
+            theme,
+            className: classes.helpHubLauncherIcon({ open: isOpen }),
+          }),
+        ])}
+        xmlns="http://www.w3.org/2000/svg"
+        viewBox="0 0 24 24"
+        fill="none"
+        strokeWidth="1.5"
+        stroke="currentColor"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      >
+        <path stroke="none" d="M0 0h24v24H0z" fill="none" />
+        <path d="M12 16v.01" />
+        <path d="M12 13a2 2 0 0 0 .914 -3.782a1.98 1.98 0 0 0 -2.414 .483" />
+      </svg>
     </div>
   );
 }