+++ packages/@dopt/react/src/context.ts
@@ -1,18 +1,7 @@
 import { Logger } from '@dopt/logger';
-import { RefObject, createContext } from 'react';
-import type {
-  Flow as APIFlow,
-  Block as APIBlock,
-  Field as APIField,
-} from '@dopt/javascript-common';
+import { MutableRefObject, createContext } from 'react';
 
-import {
-  FlowStatus,
-  Flows,
-  Blocks,
-  BlockTransitionHandler,
-  FlowIntentHandler,
-} from './types';
+import { ProviderConfig, Socket, SocketStatus } from './types';
 
 /**
  * The context stored within Dopt's provider. It is
@@ -20,16 +9,12 @@ import {
  * {@link useBlock} and {@link useFlow}.
  */
 type DoptContext = {
-  fetching: boolean;
-  flows: Flows;
-  flowBlocks: Map<APIFlow['sid'], APIBlock['uid'][]>;
-  flowStatuses: Record<APIFlow['uid'], FlowStatus>;
-  flowIntention: FlowIntentHandler;
-  blocks: Blocks;
-  blockUidBySid: Map<APIBlock['sid'], APIBlock['uid']>;
-  blockIntention: BlockTransitionHandler;
-  blockFields: Map<APIBlock['uid'], Map<APIField['sid'], APIField>>;
-  log: RefObject<Logger>;
+  socket: Socket;
+  socketStatus: SocketStatus;
+  logger: MutableRefObject<Logger>;
+  userId: ProviderConfig['userId'];
+  groupId?: ProviderConfig['groupId'];
+  apiKey: ProviderConfig['apiKey'];
 };
 
 const DoptContext = createContext<DoptContext>({} as DoptContext);