+++ apps/@examples/card/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @examples/card
 
+## 0.0.30
+
+### Patch Changes
+
+- Updated dependencies [a6cfcb0fa]
+- Updated dependencies [e027a3980]
+- Updated dependencies [b1aa72063]
+  - @dopt/react@5.0.0
+  - @dopt/react-card@10.0.0
+  - @dopt/react-users@0.2.9
+
 ## 0.0.29
 
 ### Patch Changes