+++ apps/@examples/invite-users/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @examples/invite-users
 
+## 0.0.60
+
+### Patch Changes
+
+- Updated dependencies [a6cfcb0fa]
+- Updated dependencies [e027a3980]
+- Updated dependencies [b1aa72063]
+  - @dopt/react@5.0.0
+  - @dopt/react-checklist@11.0.0
+  - @dopt/react-modal@11.0.0
+  - @dopt/react-users@0.2.9
+
 ## 0.0.59
 
 ### Patch Changes